chore: implement review tips (implement css modules, move css files closer to their components
