+++ src/components/SplashScreen.module.css
@@ -1,0 +1,7 @@
+@import "../style/fonts.css";
+
+.splashScreen {
+  height: 100dvh;
+  overflow: hidden;
+  position: relative;
+}
