--- src/components/splashScreen/style/Circle.css
@@ -1,20 +1,0 @@
-@import "../../../style/vars/borders.css";
-
-.circle-row__circle {
-  height: 106.5dvh;
-  width: 106.5dvh;
-}
-.circle-column__circle {
-  height: 62dvh;
-  width: 62dvh;
-}
-.circle-row__circle,
-.circle-column__circle {
-  z-index: 5;
-  top: 50%;
-  left: 50%;
-  border-radius: 100%;
-  border: var(--border-width) solid black;
-  position: absolute;
-  overflow: hidden;
-}
